Move domain specific value handlings to pkg/vdef package
