+++ pkg/cmd/commands/authstatus/read.go
