+++ pkg/cmd/commands/authstatus/resource.go
