+++ pkg/cmd/commands/authstatus/zz_read_gen.go
