+++ pkg/cmd/commands/disk/apply.go
