+++ pkg/cmd/commands/disk/delete.go
