+++ pkg/cmd/commands/disk/disconnect_from_server.go
