+++ pkg/cmd/commands/disk/list.go
