+++ pkg/cmd/commands/disk/list_test.go
