+++ pkg/cmd/commands/disk/resource.go
