+++ pkg/cmd/commands/disk/wait_until_ready.go
