+++ pkg/cmd/commands/disk/zz_connect_to_server_gen.go
