+++ pkg/cmd/commands/disk/zz_delete_gen.go
