+++ pkg/cmd/commands/disk/zz_disconnect_from_server_gen.go
