+++ pkg/cmd/commands/disk/zz_list_gen.go
