+++ pkg/cmd/commands/disk/zz_monitor_disk_gen.go
