+++ pkg/cmd/commands/disk/zz_read_gen.go
