+++ pkg/cmd/commands/disk/zz_update_gen.go
