+++ pkg/cmd/commands/disk/zz_wait_until_ready_gen.go
